ml/training/requirements.txt
# Requirements for train_cls.py - Advanced classifier with distillation
torch>=2.0.0
transformers>=4.35.0,<5.0.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
accelerate>=0.24.0
safetensors>=0.4.0
PyYAML>=6.0
